+++ src/components/Head.tsx
@@ -1,6 +1,7 @@
 import React from 'react';
 import NextHead from 'next/head';
 
+const isProd = process.env.NODE_ENV === 'production';
 const Head = () => {
   return (
     <NextHead>
@@ -9,7 +10,18 @@ const Head = () => {
         name="description"
         content="Pixore, an web-based editor for pixel art"
       />
-      <script async defer src="https://buttons.github.io/buttons.js"></script>
+      <script
+        async
+        src="https://www.googletagmanager.com/gtag/js?id=UA-113899183-2"
+      />
+      {isProd && (
+        <script
+          dangerouslySetInnerHTML={{
+            __html: `window.dataLayer = window.dataLayer || [];function gtag(){dataLayer.push(arguments);}gtag('js', new Date());gtag('config', 'UA-113899183-2');`,
+          }}
+        />
+      )}
+      <script async defer src="https://buttons.github.io/buttons.js" />
     </NextHead>
   );
 };