+++ src/components/Panel.tsx
@@ -51,4 +51,17 @@ const Panel: React.FC<PropTypes> = (props) => {
   );
 };
 
+const HeadlessPanel = styled.div`
+  position: absolute;
+  top: 0;
+  left: 0;
+  width: 100%;
+  height: 100%;
+  border: 2px solid #2e3440;
+  border-radius: 10px;
+  background: #3b4252;
+  overflow: hidden;
+`;
+
+export { HeadlessPanel };
 export default Panel;